N-Extras: AudioService: Allow system effect sounds to be themed
This commit checks whether there is a preexisting file in the themed
directory "/data/system/theme/audio/ui/" and if so, change the base
file paths for the sound. If the file does not exist in the theme
directory, then use the default sounds.

At the current moment, this will require a soft reboot to work.

Change-Id: I7666c2bd259443ccec442bf6059786bea3dc069e
Signed-off-by: sub77 <sub77@ymail.com>